Support Cast interfaces for deferred $ref types
Previously, a Cast using a $ref type that hadn't been defined yet would fail to find the implementation, as it couldn't yet expand the type.

We couldn't add another Cast for the $ref, as that would result in multiple implementation afterwards.

So instead we move the Casts to the end of the file, so they can see all the types.